mithril_signer

Module dependency_injection

Dependency injection module.

This module provides tools to initialize and share resources and services amongst different threads.

It takes all its inputs from the configuration which should combine inputs from:

environment

command line

configuration files

default values

The Builder ensure every service has required dependencies to build and provide services containers for each sub process.

Structs

DependenciesBuilder

The DependenciesBuilder is intended to manage Services instance creation.

SignerDependencyContainer

This structure groups all the dependencies required by the state machine.

Type Aliases

EpochServiceWrapper

EpochServiceWrapper wraps a EpochService