Module mithril_signer::metrics

metrics module. This module contains the signer metrics service and metrics server.

Structs

MetricsServer

The MetricsServer is responsible for exposing the metrics of the signer.

MetricsService

Metrics service which is responsible for recording and exposing metrics.

Constants

RUNTIME_CYCLE_SUCCESS_SINCE_STARTUP_METRIC_HELP

‘runtime_cycle_success_since_startup’ metric help

RUNTIME_CYCLE_SUCCESS_SINCE_STARTUP_METRIC_NAME

‘runtime_cycle_success_since_startup’ metric name

RUNTIME_CYCLE_TOTAL_SINCE_STARTUP_METRIC_HELP

‘runtime_cycle_total_since_startup’ metric help

RUNTIME_CYCLE_TOTAL_SINCE_STARTUP_METRIC_NAME

‘runtime_cycle_total_since_startup’ metric name

SIGNATURE_REGISTRATION_SUCCESS_LAST_EPOCH_METRIC_HELP

‘signature_registration_success_last_epoch’ metric help

SIGNATURE_REGISTRATION_SUCCESS_LAST_EPOCH_METRIC_NAME

‘signature_registration_success_last_epoch’ metric name

SIGNATURE_REGISTRATION_SUCCESS_SINCE_STARTUP_METRIC_HELP

‘signature_registration_success_since_startup’ metric help

SIGNATURE_REGISTRATION_SUCCESS_SINCE_STARTUP_METRIC_NAME

‘signature_registration_success_since_startup’ metric name

SIGNATURE_REGISTRATION_TOTAL_SINCE_STARTUP_METRIC_HELP

‘signature_registration_total_since_startup’ metric help

SIGNATURE_REGISTRATION_TOTAL_SINCE_STARTUP_METRIC_NAME

‘signature_registration_total_since_startup’ metric name

SIGNER_REGISTRATION_SUCCESS_LAST_EPOCH_METRIC_HELP

‘signer_registration_success_last_epoch’ metric help

SIGNER_REGISTRATION_SUCCESS_LAST_EPOCH_METRIC_NAME

‘signer_registration_success_last_epoch’ metric name

SIGNER_REGISTRATION_SUCCESS_SINCE_STARTUP_METRIC_HELP

‘signer_registration_success_since_startup’ metric help

SIGNER_REGISTRATION_SUCCESS_SINCE_STARTUP_METRIC_NAME

‘signer_registration_success_since_startup’ metric name

SIGNER_REGISTRATION_TOTAL_SINCE_STARTUP_METRIC_HELP

‘signer_registration_total_since_startup’ metric help

SIGNER_REGISTRATION_TOTAL_SINCE_STARTUP_METRIC_NAME

‘signer_registration_total_since_startup’ metric name