Module mithril_signer::services

Services

This module regroups services. Services are adapters in charge of the different bounded contexts of the application:

Aggregator Client: communicate with the Aggregator

Cardano Transactions: handle Cardano transactions (import, preload, etc.)

Single Signer: create single signatures

Upkeep: perform maintenance tasks

Each service is defined by a public API (a trait) that is used in the controllers (runtimes).

Structs

AggregatorHTTPClient

AggregatorHTTPClient is a http client for an aggregator

CardanoTransactionsImporter

Import and store CardanoTransaction.

CardanoTransactionsPreloaderActivationSigner

CardanoTransactionsPreloaderActivationSigner

MithrilEpochService

Implementation of the epoch service.

MithrilProtocolInitializerBuilder

This is responsible for creating new instances of ProtocolInitializer.

MithrilSingleSigner

Implementation of the SingleSigner.

SignerCertifierService

Implementation of the Certifier Service for the Mithril Signer.

SignerSignableSeedBuilder

SignableSeedBuilder signer implementation

SignerSignedEntityConfigProvider

Simple wrapper to the EpochService to implement the SignedEntityConfigProvider trait.

SignerUpkeepService

Implementation of the upkeep service for the signer.

TransactionsImporterByChunk

A decorator of TransactionsImporter that does the import by chunks

TransactionsImporterWithPruner

A decorator of TransactionsImporter that prunes the transactions older than a given number of blocks after running the import.

TransactionsImporterWithVacuum

A decorator of TransactionsImporter that vacuums the database after running the import.

Enums

AggregatorClientError

Error structure for the Aggregator Client.

EpochServiceError

Errors dedicated to the EpochService.

SingleSignerError

SingleSigner error structure.

Traits

AggregatorClient

Trait for mocking and testing a AggregatorClient

CertifierService

Certifier Service

EpochPruningTask

Define the task responsible for pruning a datasource below a certain epoch threshold.

EpochService

Service that aggregates all data that don’t change in a given epoch.

HighestTransactionBlockNumberGetter

Trait to get the highest transaction block number

SignaturePublisher

Publishes computed single signatures to a third party.

SignedBeaconStore

Trait to store beacons that have been signed in order to avoid signing them twice.

SignedEntityConfigProvider

Trait to provide the current signed entity configuration that can change over time.

SingleSigner

The SingleSigner is the structure responsible for issuing SingleSignatures.

TransactionPruner

Cardano transactions pruner

TransactionStore

Cardano transactions store

UpkeepService

Define the service responsible for the upkeep of the application.