mithril_signer::services

Trait SignedBeaconStore

pub trait SignedBeaconStore: Sync + Send {
    // Required methods
    fn filter_out_already_signed_entities<'life0, 'async_trait>(
        &'life0 self,
        entities: Vec<SignedEntityType>,
    ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntityType>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn mark_beacon_as_signed<'life0, 'life1, 'async_trait>(
        &'life0 self,
        entity: &'life1 BeaconToSign,
    ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait to store beacons that have been signed in order to avoid signing them twice.

Required Methods

fn filter_out_already_signed_entities<'life0, 'async_trait>( &'life0 self, entities: Vec<SignedEntityType>, ) -> Pin<Box<dyn Future<Output = StdResult<Vec<SignedEntityType>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Filter out already signed entities from a list of signed entities.

fn mark_beacon_as_signed<'life0, 'life1, 'async_trait>( &'life0 self, entity: &'life1 BeaconToSign, ) -> Pin<Box<dyn Future<Output = StdResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Mark a beacon as signed.