mithril_signer
Struct Configuration

pub struct Configuration {

    pub cardano_cli_path: PathBuf,
    pub cardano_node_socket_path: PathBuf,
    pub network: String,
    pub network_magic: Option<u64>,
    pub network_security_parameter: BlockNumber,
    pub preload_security_parameter: BlockNumber,
    pub aggregator_endpoint: String,
    pub relay_endpoint: Option<String>,
    pub party_id: Option<PartyId>,
    pub run_interval: u64,
    pub db_directory: PathBuf,
    pub data_stores_directory: PathBuf,
    pub store_retention_limit: Option<usize>,
    pub kes_secret_key_path: Option<PathBuf>,
    pub operational_certificate_path: Option<PathBuf>,
    pub disable_digests_cache: bool,
    pub reset_digests_cache: bool,
    pub era_reader_adapter_type: EraReaderAdapterType,
    pub era_reader_adapter_params: Option<String>,
    pub enable_metrics_server: bool,
    pub metrics_server_ip: String,
    pub metrics_server_port: u16,
    pub allow_unparsable_block: bool,
    pub enable_transaction_pruning: bool,
    pub transactions_import_block_chunk_size: BlockNumber,
    pub cardano_transactions_block_streamer_max_roll_forwards_per_poll: usize,
    pub preloading_refresh_interval_in_seconds: u64,

}

Client configuration

Fields
cardano_cli_path: PathBuf

Cardano CLI tool path

cardano_node_socket_path: PathBuf

Path of the socket used by the Cardano CLI tool to communicate with the Cardano node

network: String

Cardano network

network_magic: Option<u64>

Cardano Network Magic number useful for TestNet & DevNet

network_security_parameter: BlockNumber

Also known as k, it defines the number of blocks that are required for the blockchain to be considered final, preventing any further rollback [default: 2160].

preload_security_parameter: BlockNumber

Blocks offset, from the tip of the chain, to exclude during the cardano transactions preload [default: 3000].

aggregator_endpoint: String

Aggregator endpoint

relay_endpoint: Option<String>

Relay endpoint

party_id: Option<PartyId>

Party Id

run_interval: u64

Run Interval

db_directory: PathBuf

Directory to snapshot

data_stores_directory: PathBuf

Directory to store signer data (Stakes, Protocol initializers, …)

store_retention_limit: Option<usize>

Store retention limit. If set to None, no limit will be set.

kes_secret_key_path: Option<PathBuf>

File path to the KES secret key of the pool

operational_certificate_path: Option<PathBuf>

File path to the operational certificate of the pool

disable_digests_cache: bool

Disable immutables digests cache.

reset_digests_cache: bool

If set the existing immutables digests cache will be reset.

Will be ignored if set in conjunction with disable_digests_cache.

era_reader_adapter_type: EraReaderAdapterType

Era reader adapter type

era_reader_adapter_params: Option<String>

Era reader adapter parameters

enable_metrics_server: bool

Enable metrics server (Prometheus endpoint on /metrics).

metrics_server_ip: String

Metrics HTTP Server IP.

metrics_server_port: u16

Metrics HTTP Server listening port.

allow_unparsable_block: bool

If set no error is returned in case of unparsable block and an error log is written instead.

Will be ignored on (pre)production networks.

enable_transaction_pruning: bool

If set, the signer will prune the cardano transactions in database older than the network_security_parameter blocks after each import [default: true].

transactions_import_block_chunk_size: BlockNumber

Chunk size for importing transactions, combined with transaction pruning it reduces the storage footprint of the signer by reducing the number of transactions stored on disk at any given time.

cardano_transactions_block_streamer_max_roll_forwards_per_poll: usize

The maximum number of roll forwards during a poll of the block streamer when importing transactions.

preloading_refresh_interval_in_seconds: u64

Preloading refresh interval in seconds

Implementations

impl Configuration

pub fn get_network(&self) -> StdResult<CardanoNetwork>

Return the CardanoNetwork value from the configuration.

pub fn get_sqlite_file(&self, sqlite_file_name: &str) -> StdResult<PathBuf>

Create the SQL store directory if not exist and return the path of the SQLite3 file.

pub fn build_era_reader_adapter( &self, chain_observer: Arc<dyn ChainObserver>, ) -> StdResult<Arc<dyn EraReaderAdapter>>

Create era reader adapter from configuration settings.