Trait mithril_signer::ServiceBuilder

pub trait ServiceBuilder {
    fn build(&self) -> Result<SignerServices, Box<dyn StdError>>;
}

The ServiceBuilder is intended to manage Services instance creation. The goal of this is to put all this code out of the way of business code.

Required Methods

fn build(&self) -> Result<SignerServices, Box<dyn StdError>>

Create a SignerService instance.