Crate mithril_signer

Mithril Signer crate documentation

This crate is used by Cardano nodes to participate to Mithril signatures. It proposes tools to communicate with Mithril aggregators and to issue Single Signatures. See the Mithril documentation for more information on how it works.

Re-exports

pub use metrics::*;

Modules

database

database module. This module contains the entities definition tied with database representation with their associated providers.

metrics

metrics module. This module contains the signer metrics service and metrics server.

Structs

AggregatorHTTPClient

AggregatorHTTPClient is a http client for an aggregator

CardanoTransactionsImporter

Import and store CardanoTransaction.

CardanoTransactionsPreloaderActivationSigner

CardanoTransactionsPreloaderActivationSigner

Configuration

Client configuration

DefaultConfiguration

Default configuration with all the default values for configurations.

FromEpochSettingsAdapter

Adapter to convert EpochSettingsMessage to EpochSettings.

FromPendingCertificateMessageAdapter

Adapter to turn CertificatePendingMessage instances into CertificatePending.

MithrilProtocolInitializerBuilder

This is responsible of creating new instances of ProtocolInitializer.

MithrilSingleSigner

Implementation of the SingleSigner.

ProductionServiceBuilder

Create a SignerService instance for Production environment.

ProtocolInitializerStore

Implementation of the ProtocolInitializerStorer

SignerRunner

Controller methods for the Signer’s state machine.

SignerServices

This structure groups all the services required by the state machine.

SignerUpkeepService

Implementation of the upkeep service for the signer.

StateMachine

The state machine is responsible of the execution of the signer automate.

ToRegisterSignerMessageAdapter

Adapter to create RegisterSignerMessage from Signer instance.

TransactionsImporterByChunk

A decorator of TransactionsImporter that does the import by chunks

TransactionsImporterWithPruner

A decorator of TransactionsImporter that prunes the transactions older than a given number of blocks after running the import.

TransactionsImporterWithVacuum

A decorator of TransactionsImporter that vacuums the database after running the import.

Enums

AggregatorClientError

Error structure for the Aggregator Client.

RunnerError

This type represents the errors thrown from the Runner.

RuntimeError

RuntimeError Error kinds tied to their faith in the state machine.

SignerState

Different possible states of the state machine.

SingleSignerError

SingleSigner error structure.

Traits

AggregatorClient

Trait for mocking and testing a AggregatorClient

HighestTransactionBlockNumberGetter

Trait to get the highest transaction block number

ProtocolInitializerStorer

Store the ProtocolInitializer used for each Epoch. This is useful because protocol parameters and stake distribution change over time.

Runner

This trait is mainly intended for mocking.

ServiceBuilder

The ServiceBuilder is intended to manage Services instance creation. The goal of this is to put all this code out of the way of business code.

SingleSigner

The SingleSigner is the structure responsible of issuing SingleSignatures.

TransactionPruner

Cardano transactions pruner

TransactionStore

Cardano transactions store

UpkeepService

Define the service responsible for the upkeep of the application.