Crate mithril_signer

Mithril Signer crate documentation

This crate is used by Cardano nodes to participate to Mithril signatures. It proposes tools to communicate with Mithril aggregators and to issue Single Signatures. See the Mithril documentation for more information on how it works.

Re-exports

pub use entities::RegisteredSigners;

pub use metrics::*;

Modules

database

database module. This module contains the entities definition tied with database representation with their associated providers.

dependency_injection

Dependency injection module.

entities

Entities module

metrics

metrics module. This module contains the signer metrics service and metrics server.

services

Services

store

Alternative storage backends when relational database capabilities are not needed.

Structs

Configuration

Client configuration

DatabaseCommand

Database tools

DefaultConfiguration

Default configuration with all the default values for configurations.

FromEpochSettingsAdapter

Adapter to convert EpochSettingsMessage to RegisteredSigners.

MigrateCommand

Migrate command

SignerRunner

Controller methods for the Signer’s state machine.

StateMachine

The state machine is responsible of the execution of the signer automate.

ToRegisterSignerMessageAdapter

Adapter to create RegisterSignerMessage from Signer instance.

Enums

DatabaseSubCommand

Database subcommands

RunnerError

This type represents the errors thrown from the Runner.

RuntimeError

RuntimeError Error kinds tied to their faith in the state machine.

SignerState

Different possible states of the state machine.

Traits

Runner

This trait is mainly intended for mocking.