Crate mithril_signer

Mithril Signer crate documentation

This crate is used by Cardano nodes to participate to Mithril signatures. It proposes tools to communicate with Mithril aggregators and to issue Single Signatures. See the Mithril documentation for more information on how it works.

Modules

database

database module. This module contains the entities definition tied with database representation with their associated providers.

Structs

AggregatorHTTPClient

AggregatorHTTPClient is a http client for an aggregator

Configuration

Client configuration

DefaultConfiguration

Default configuration with all the default values for configurations.

FromEpochSettingsAdapter

Adapter to convert EpochSettingsMessage to EpochSettings.

FromPendingCertificateMessageAdapter

Adapter to turn CertificatePendingMessage instances into CertificatePending.

MithrilProtocolInitializerBuilder

This is responsible of creating new instances of ProtocolInitializer.

MithrilSingleSigner

Implementation of the SingleSigner.

ProductionServiceBuilder

Create a SignerService instance for Production environment.

ProtocolInitializerStore

Implementation of the ProtocolInitializerStorer

SignerRunner

Controller methods for the Signer’s state machine.

SignerServices

This structure groups all the services required by the state machine.

StateMachine

The state machine is responsible of the execution of the signer automate.

ToRegisterSignerMessageAdapter

Adapter to create RegisterSignerMessage from Signer instance.

Enums

AggregatorClientError

Error structure for the Aggregator Client.

RunnerError

This type represents the errors thrown from the Runner.

RuntimeError

RuntimeError Error kinds tied to their faith in the state machine.

SignerState

Different possible states of the state machine.

SingleSignerError

SingleSigner error structure.

Traits

AggregatorClient

Trait for mocking and testing a AggregatorClient

ProtocolInitializerStorer

Store the ProtocolInitializer used for each Epoch. This is useful because protocol parameters and stake distribution change over time.

Runner

This trait is mainly intended for mocking.

ServiceBuilder

The ServiceBuilder is intended to manage Services instance creation. The goal of this is to put all this code out of the way of business code.

SingleSigner

The SingleSigner is the structure responsible of issuing SingleSignatures.