mithril_signer

Module services

Services

This module regroups services. Services are adapters in charge of the different bounded contexts of the application:

Aggregator Client: communicate with the Aggregator

Cardano Transactions: handle Cardano transactions (import, preload, etc.)

Single Signer: create single signatures

Upkeep: perform maintenance tasks

Each service is defined by a public API (a trait) that is used in the controllers (runtimes).

Structs

CardanoTransactionsPreloaderActivationSigner

CardanoTransactionsPreloaderActivationSigner

ChainDataImporterWithVacuum

A decorator of [ChainDataImporter] that vacuums the database after running the import.

MithrilEpochService

Implementation of the epoch service.

MithrilProtocolInitializerBuilder

This is responsible for creating new instances of ProtocolInitializer.

MithrilSingleSigner

Implementation of the SingleSigner.

SignaturePublishRetryPolicy

Policy for retrying signature publishing

SignaturePublisherDelayer

A decorator of SignaturePublisher that publishes right away on a first publisher and with a delay on the second publisher.

SignaturePublisherDmq

DMQ implementation of the SignaturePublisher trait.

SignaturePublisherNoop

A no-op implementation of the SignaturePublisher trait. This implementation performs no action when a signature is published.

SignaturePublisherRetrier

A decorator of SignaturePublisher that retries the publishing of signatures in case of failure

SignerCertifierService

Implementation of the Certifier Service for the Mithril Signer.

SignerChainDataImporter

Services responsible for importing chain data into the signer data stores.

SignerSignableSeedBuilder

SignableSeedBuilder signer implementation

SignerSignedEntityConfigProvider

Simple wrapper to the EpochService to implement the SignedEntityConfigProvider trait.

SignerUpkeepService

Implementation of the upkeep service for the signer.

Enums

EpochServiceError

Errors dedicated to the EpochService.

SingleSignerError

SingleSigner error structure.

Traits

CertifierService

Certifier Service

EpochPruningTask

Define the task responsible for pruning a datasource below a certain epoch threshold.

EpochService

Service that aggregates all data that don’t change in a given epoch.

SignaturePublisher

Publishes computed single signature to a third party.

SignedBeaconStore

Trait to store beacons that have been signed in order to avoid signing them twice.

SignedEntityConfigProvider

Trait to provide the current signed entity configuration that can change over time.

SignerRegistrationPublisher

Publishes a signer registration to a third party.

SignersRegistrationRetriever

Service responsible for retrieving the signer’s registration from the mithril network

SingleSigner

The SingleSigner is the structure responsible for issuing SingleSignature.

UpkeepService

Define the service responsible for the upkeep of the application.